% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\docType{data}
\name{dataXY_df}
\alias{dataXY_df}
\title{Terra satellite data (X,Y) for running the xgboost model .}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 10148 rows and 10 columns.
}
\usage{
dataXY_df
}
\description{
Data.table, contains 9 features, and about 10,000 observations
}
\references{
\url{http://doi.org/10.5281/zenodo.3334713}
}
\keyword{Terra}
