% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHAP_funcs.R
\name{label.feature}
\alias{label.feature}
\title{helper function to modify labels for features under plotting}
\usage{
label.feature(x)
}
\arguments{
\item{x}{variable names}
}
\value{
a character, e.g. "date", "Time Trend", etc.
}
\description{
If a list is created in the environment named \strong{new_labels} (\code{!is.null(new_labels}),
the plots will use that list to replace default list of labels \code{\link{labels_within_package}}.
}
