% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\docType{data}
\name{new_labels}
\alias{new_labels}
\title{new_labels: a place holder, if supplied as a list, it offers user to rename labels}
\format{An object of class \code{NULL} of length 0.}
\usage{
new_labels
}
\description{
new_labels: a place holder, if supplied as a list, it offers user to rename labels
}
\keyword{Labels}
