% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHAP_funcs.R
\name{plot.label}
\alias{plot.label}
\title{sub-function to revise label for each feature.}
\usage{
\method{plot}{label}(plot1, show_feature)
}
\arguments{
\item{plot1}{ggplot2 object}

\item{show_feature}{feature to plot}
}
\description{
This function further fine-tune the format of each feature
}
