% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHAP_funcs.R
\name{plot.shap.core}
\alias{plot.shap.core}
\title{the basic SHAP core, accept data with only 'show_feature' as variable.}
\usage{
\method{plot}{shap.core}(show_feature, data0)
}
\description{
the core function to plot SHAP for one feature.
}
