% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHAP_funcs.R
\name{plot.shap.dependence.color}
\alias{plot.shap.dependence.color}
\title{SHAP dependence plot with marginal histogram, colored by a selcted feature value}
\usage{
\method{plot}{shap.dependence.color}(data_long, x, y_shap, color_feature)
}
\arguments{
\item{data_long}{the long format SHAP values}

\item{x}{which feature to show on x axis}

\item{y_shap}{which shap values to show on y axis}

\item{color_feature}{which feature value to use for coloring}
}
\description{
SHAP dependence plot with marginal histogram, colored by a selcted feature value
}
\examples{
data("iris")
X1 = as.matrix(iris[,-5])
mod1 = xgboost(
  data = X1, label = iris$Species, gamma = 0, eta = 1, lambda = 0,nrounds = 1, verbose = F)

# shap.values() has the SHAP data matrix and ranked features by mean|SHAP|
shap_values <- shap.values(mod1, X1)
shap_values$mean_shap_score

# shap.prep() returns the long-format SHAP data
shap_long <- shap.prep(shap_values, X1)


# **SHAP dependence plot**
plot.shap.dependence.color(shap_long, x="Petal.Length",
                           y_shap = "Petal.Length", color_feature = "Petal.Width")

# **SHAP force plot**
plot_data <- shap.stack.data(shap_contrib = shap_values$shap_score, n_groups = 2)
plot.shap.force_plot(plot_data)
plot.shap.force_plot_bygroup(plot_data)
}
