% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHAP_funcs.R
\name{plot.shap.dependence}
\alias{plot.shap.dependence}
\title{SHAP dependence plot with marginal histogram.}
\usage{
\method{plot}{shap.dependence}(data_long, show_feature, dilute = FALSE,
  customize_label = FALSE)
}
\arguments{
\item{data_long}{the long format SHAP values}

\item{show_feature}{which feature to show}

\item{dilute}{a number or logical, dafault to TRUE, will plot \code{nrow(data_long)/dilute} data.
for example, if dilute = 5 will plot 1/5 of the data.}
}
\value{
ggplot2 object
}
\description{
\code{plot.shap.dependence} is not colored by another variable,
\code{\link{plot.shap.dependence.color}} is the with color version.
Dependence plot is very easy to make if you have the SHAP values dataset
It is not necessary to start with the long-format data, but since I used that
for the summary plot, I just continue to use the long dataset
}
\examples{
library(SHAPforxgboost)
# **SHAP dependence plot**
plot.shap.dependence.color(data_long = shap_long_iris, x="Petal.Length",
                           y = "Petal.Width", color_feature = "Petal.Width")
# the without color version, just plot SHAP value against feature value:
plot.shap.dependence(data_long = shap_long_iris, "Petal.Length")

}
