% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHAP_funcs.R
\name{plot.shap.force_plot}
\alias{plot.shap.force_plot}
\title{make the stack plot, optional to zoom in at certain x or certain cluster.}
\usage{
\method{plot}{shap.force_plot}(shapobs, id = "id",
  zoom_in_location = NULL, y_parent_limit = NULL,
  y_zoomin_limit = NULL, zoom_in = TRUE, zoom_in_group = NULL)
}
\description{
make the stack plot, optional to zoom in at certain x or certain cluster.
}
\examples{

# **SHAP force plot**
plot_data <- shap.stack.data(shap_contrib = shap_values_iris, n_groups = 2)
plot.shap.force_plot(plot_data)
plot.shap.force_plot_bygroup(plot_data)
}
