% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHAP_funcs.R
\name{plot.shap.force_plot_bygroup}
\alias{plot.shap.force_plot_bygroup}
\title{make the stack plot, optional to zoom in at certain x or certain cluster.}
\usage{
\method{plot}{shap.force_plot_bygroup}(shapobs, id = "id",
  y_parent_limit = NULL)
}
\description{
Zoomed in plot: one plot of every group of clustered observations.
}
\examples{

# **SHAP force plot**
plot_data <- shap.stack.data(shap_contrib = shap_values_iris, n_groups = 2)
plot.shap.force_plot(plot_data)
plot.shap.force_plot_bygroup(plot_data)
}
