% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHAP_funcs.R
\name{plot.shap.summary}
\alias{plot.shap.summary}
\title{SHAP summary plot
The summary plot (sina plot) using a long-format data of SHAP values.}
\usage{
\method{plot}{shap.summary}(data_long, x_bound_given = NULL,
  dilute = F, scientific = F)
}
\arguments{
\item{x_bound_given}{in case need to limit x_axis_limit}

\item{dilute}{by default print all data, but that would be slow}

\item{scientific}{default to F, "%.1e", show the mean|SHAP| in scientific or not
if nobs > 10,000. if dilute, it will use at most 1000 observations}

\item{data_long:}{a long format data of SHAP values which looks like:
variable value rfvalue stdfvalue mean_value
1:   dayint -0.23   11012   0.00000       0.14
2:   dayint -0.20   11012   0.00000       0.14}

\item{dilute}{dafault to F, if dilute, will plot at most 1,000 points per feature.}
}
\description{
SHAP summary plot
The summary plot (sina plot) using a long-format data of SHAP values.
}
\examples{

data("iris")
X1 = as.matrix(iris[,-5])
mod1 = xgboost(
  data = X1, label = iris$Species, gamma = 0, eta = 1, lambda = 0,nrounds = 1, verbose = F)

# shap.values() has the SHAP data matrix and ranked features by mean|SHAP|
shap_values <- shap.values(mod1, X1)
shap_values$mean_shap_score

# shap.prep() returns the long-format SHAP data
shap_long <- shap.prep(shap_values, X1)

# **SHAP summary plot**
plot.shap.summary(shap_long)

# Alternatives:
# option 1: from the xgboost model
plot.shap.summary.wrap1(mod1, X1, top_n = 3)

# option 2: supply a self-made SHAP values dataset (e.g. sometimes as output from cross-validation)
plot.shap.summary.wrap2(shap_score = shap_values$shap_score, X1, top_n = 3)
}
