% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_plots.R
\name{scatter.plot.diagonal}
\alias{scatter.plot.diagonal}
\title{make customized scatter plot with diagonal line and R2 printed.}
\usage{
scatter.plot.diagonal(data, x, y, dilute = F, add_abline = F,
  add_hist = T)
}
\arguments{
\item{data}{dataset}

\item{x}{x}

\item{y}{y}

\item{dilute}{a number or logical, dafault to TRUE, will plot \code{nrow(data_long)/dilute} data. For example, if dilute = 5 will plot 1/5 of the data.}

\item{add_abline}{default to FALSE, add a diagonal line}

\item{add_hist}{optional to add marginal histogram using \code{ggExtra::ggMarginal} but notice
if add histogram, what is returned is no longer a ggplot2 object}
}
\value{
ggplot2 object if \code{add_hist = FALSE}
}
\description{
make customized scatter plot with diagonal line and R2 printed.
}
\examples{
scatter.plot.diagonal(data = iris, x = "Sepal.Length", y = "Petal.Length")

}
