% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHAP_funcs.R
\name{shap.plot.core}
\alias{shap.plot.core}
\title{the basic SHAP core, accept data with only 'show_feature' as variable.}
\usage{
shap.plot.core(show_feature, data0)
}
\arguments{
\item{show_feature}{the feature to plot}

\item{data0}{part of the data_long that contains the feature}
}
\description{
the core function to plot SHAP for one feature.
}
