% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHAP_funcs.R
\name{shap.plot.force_plot_bygroup}
\alias{shap.plot.force_plot_bygroup}
\title{Make the stack plot, optional to zoom in at certain x or certain cluster}
\usage{
shap.plot.force_plot_bygroup(shapobs, id = "sorted_id", y_parent_limit = NULL)
}
\arguments{
\item{shapobs}{The dataset obtained by \code{shap.prep.stack.data}.}

\item{id}{the id variable.}

\item{y_parent_limit}{set y-axis limits.}
}
\description{
A collective display of zoom-in plots: one plot for every group of the
clustered observations.
}
\examples{

# **SHAP force plot**
plot_data <- shap.prep.stack.data(shap_contrib = shap_values_iris,
                                  n_groups = 4)
shap.plot.force_plot(plot_data)
shap.plot.force_plot(plot_data,  zoom_in_group = 2)

# plot all the clusters:
shap.plot.force_plot_bygroup(plot_data)
}
