% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\docType{data}
\name{shap_int_iris}
\alias{shap_int_iris}
\title{The interaction effect SHAP values example using iris dataset.}
\format{An object of class \code{array} of dimension 150 x 5 x 5.}
\usage{
shap_int_iris
}
\description{
The interaction effect SHAP values example using iris dataset.
}
\keyword{iris}
