% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\docType{data}
\name{shap_long_iris}
\alias{shap_long_iris}
\title{The long-format SHAP values example using iris dataset.}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 600 rows and 6 columns.}
\usage{
shap_long_iris
}
\description{
The long-format SHAP values example using iris dataset.
}
\keyword{iris}
