% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\docType{data}
\name{shap_score}
\alias{shap_score}
\title{SHAP values example from dataXY_df .}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 10148 rows and 9 columns.
}
\usage{
shap_score
}
\description{
SHAP values example from dataXY_df .
}
\references{
\url{http://doi.org/10.5281/zenodo.3334713}
}
\keyword{Terra}
