% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHAP_funcs.R
\name{label.feature}
\alias{label.feature}
\title{helper function to modify variables names.}
\usage{
label.feature(x, labs)
}
\arguments{
\item{x}{feature names}
}
\description{
helper function to modify variables names.
}
