% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHAP_funcs.R
\name{shap.plot.force_plot_bygroup}
\alias{shap.plot.force_plot_bygroup}
\title{make the stack plot, optional to zoom in at certain x or certain cluster.}
\usage{
shap.plot.force_plot_bygroup(shapobs, id = "id", y_parent_limit = NULL)
}
\arguments{
\item{shapobs}{The dataset obtained by \code{shap.prep.stack.data}.}

\item{id}{the id variable.}

\item{y_parent_limit}{set y axis limits.}
}
\description{
A collective display of zoom in plot: one plot of every group of clustered observations.
}
\examples{

# **SHAP force plot**
plot_data <- shap.prep.stack.data(shap_contrib = shap_values_iris,
                                  n_groups = 2)
shap.plot.force_plot(plot_data)
shap.plot.force_plot_bygroup(plot_data)
}
