% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strongest_interaction}
\alias{strongest_interaction}
\title{Finds variable with presumably strongest interaction effect}
\usage{
strongest_interaction(X0, Xlong)
}
\description{
Finds variable with presumably strongest interaction effect
}
